Fixed 500 errors caused by simplejson if it was installed
Flask uses simplejson instead of the default json module as its serializing backend if it is available.
However, the custom JSON Encoder used in Silence extended the default json module encoder. This caused errors if simplejson was available,
because Silence's encoder wouldn't accept parameters that simplejson's would.

This has been fixed by importing flask's json module instead of the builtin json, which is an alias for whatever backend it is using.